Moving marshalling and schema to config package
